requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.0.1
pytesseract==0.3.10
